Fix private lib dependencies that produce LINK_ONLY targets
At this moment, trying to wrap the following code will fail:

```
add_library(bar OBJECT bar.c)
add_library(foo OBJECT foo.c)
target_link_libraries(foo PRIVATE bar)
```

This is because when analyzing dependencies of `foo`, the `bar`
library will be wrapped as `$<LINK_ONLY:bar>`. It is an expected
behavior of CMake. But we try to generate a `-config.cmake` while
expanding generator macros, and the `LINK_ONLY` macro can only
be expanded in the linker context.

Thus, the fisr thing that comes to mind is to skip pulling private
libraries. It fixes code generation in many cases, and doesn't affect
the code that works right now.